+++ test_bin/src/publish_hold_workflows.rs
@@ -1,15 +1,16 @@
 
-use crate::*;
-
 use holochain_core_types::cas::content::Address;
 use holochain_net_connection::{
     json_protocol::{
         ConnectData, DhtData, DhtMetaData, FetchDhtData, FetchDhtMetaData, JsonProtocol, MessageData,
-        TrackDnaData, HandleListResultData, GetListData,
+        TrackDnaData, HandleListResultData, GetListData, FailureResultData,
     },
     net_connection::NetSend,
     NetResult,
 };
+use p2p_node::P2pNode;
+use constants::*;
+use basic_workflows::setup_normal;
 
 /// Macro for transforming a type check into a predicate
 macro_rules! one_is {
@@ -25,7 +26,7 @@ macro_rules! one_is {
 
 // CONSTS
 static PUBLISH_LIST_ID: &'static str = "publish_list_id_1";
-static FETCH_ENTRY_1_ID: &'static str = "fetch_entry_1_id";
+
 
 /// Test the following workflow after normal setup:
 /// sequenceDiagram
@@ -36,7 +37,7 @@ static FETCH_ENTRY_1_ID: &'static str = "fetch_entry_1_id";
 /// net->>a: HandleFetchData(xyz_addr)
 /// a->>net: HandleFetchDataResult(xyz)
 #[cfg_attr(tarpaulin, skip)]
-fn empty_publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_connect: bool) -> NetResult<()> {
+pub fn empty_publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_connect: bool) -> NetResult<()> {
     // Setup
     println!("Testing: empty_publish_data_list_test()");
     setup_normal(alex, billy, can_connect)?;
@@ -81,6 +82,17 @@ fn empty_publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_con
             .into(),
     )?;
 
+    // Alex sends that data back to the network
+    alex.send(
+        JsonProtocol::FailureResult(FailureResultData {
+            request_id: FETCH_ENTRY_1_ID.to_string(),
+            dna_address: example_dna_address(),
+            to_agent_id: BILLY_AGENT_ID.to_string(),
+            error_info: json!("does not have data"),
+        })
+            .into(),
+    )?;
+
     let result = billy.wait(Box::new(one_is!(JsonProtocol::FailureResult(_))))?;
     println!("got result: {:?}", result);
 
@@ -91,7 +103,7 @@ fn empty_publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_con
 
 
 #[cfg_attr(tarpaulin, skip)]
-fn publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_connect: bool) -> NetResult<()> {
+pub fn publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_connect: bool) -> NetResult<()> {
     // Setup
     println!("Testing: publish_data_list_test()");
     setup_normal(alex, billy, can_connect)?;
@@ -106,25 +118,7 @@ fn publish_data_list_test(alex: &mut P2pNode, billy: &mut P2pNode, can_connect: 
             .into(),
     )?;
 
-//    // Mock a HandleFetchDhtData
-//    {
-//        let server_msg = JsonProtocol::HandleFetchDhtData(FetchDhtData {
-//            request_id: "get_dht_data_1".to_string(),
-//            dna_address: example_dna_address(),
-//            pub requester_agent_id: String,
-//        })
-//            .into();
-//
-//        let server = get_server(alex);
-//        server.mock_send_one(
-//            example_dna_address(),
-//            ALEX_AGENT_ID,
-//        data: Protocol,
-//        ).expect("mock_send_one");
-//
-//    }
-
-    // billy asks for unpublished data.
+    // billy asks for published data.
     // Shoud get a failure response
     billy.send(
         JsonProtocol::FetchDhtData(FetchDhtData {
